corrections due to atomic host path to follow specification guidelines for min/max spec cases
